+++ src/index.js
@@ -3,6 +3,7 @@ import './assets/css/chosen.css'
 import './assets/css/main.css'
 import 'chosen-js'
 import OkChart from './OkChart'
+
 ;(function(win, doc) {
   const loadGoogleApi = function(w, d, s, g, js, fs) {
     g = w.gapi || (w.gapi = {})
